+++ lib/sidekiq_glass.rb
@@ -1,8 +1,0 @@
-require 'sidekiq'
-require 'null_logger'
-
-base_path = File.dirname(__FILE__) + '/sidekiq_glass'
-
-%w( worker timeout version ).each do |file|
-  require File.join(base_path, file)
-end