+++ spec/lib/sidekiq_glass/version_spec.rb
@@ -1,5 +1,0 @@
-require 'spec_helper'
-
-RSpec.describe SidekiqGlass do
-  pending
-end