+++ spec/lib/worker_glass/version_spec.rb
@@ -1,0 +1,5 @@
+require 'spec_helper'
+
+RSpec.describe WorkerGlass do
+  it { expect { described_class::VERSION }.to_not raise_error }
+end