+++ CHANGELOG.md
@@ -1,4 +1,11 @@
-# SidekiqGlass changelog
+# WorkerGlass changelog
+
+## 0.2.0
+- Renamed from SidekiqGlass to WorkerGlass because if now does not require Sidekiq anymore to work. It will provide additional worker functionalities to any type of backend processing layer as long as it has a #perform method.
+- Switched to Mongoid like feature providing by including/prepending functionalities
+- WorkerGlass no longer required
+- WorkerGlass::Timeout
+- WorkerGlass::Reentrancy
 
 ## 0.1.4